💫 (#28) - Refactor admin service and DTOs
