--- src/admin/dto/response/get-created-by-requestor.dto.ts
@@ -1,11 +1,0 @@
-export interface GetCreatedByRequestorDto {
-  count: number;
-  accounts: Account[];
-}
-
-interface Account {
-  uuid: string;
-  uid: string;
-  name: string;
-  emoji: string;
-}
